ampart: remove package and installation script
The ampart package included in EmuELEC is very old and out-of-date,
and the version is poorly documented. It does not support safe DTB
modification modes like the new versions, either.

Bumping it to the post-v1.0 versions is a possible choice, but that
will render the installation script useless, which I also don't
have time to update due to great change in ampart's API and func-
tionality introcuted after v1.0. A re-written GUI installer that
supports user-defined partition layout would be better and grant
users much more freedom, but I also don't have time to implement
that for now.

As the author of ampart I don't want the package to be stucked
in a limbo state, so for now please remove the pacakge.

It will be re-added by myself if I have time to finish a new GUI
installer that depends on it. For now, after it is removed, if any
user still want to use ampart, they can always just download the
pre-compiled aarch64 releases from ampart's release page:
https://github.com/7Ji/ampart/releases
And I believe power users can easily do this.